# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hirevec API
  version: 1.0.0
  description: >
    Job-matching API.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FailData:
      type: object
      additionalProperties:
        type: string
      example:
        hide_reacted: "must be one of: false, true"
    FailResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [fail]
        data:
          $ref: "#/components/schemas/FailData"
...
paths:
...
  /v1/auth/token:
    post:
      operationId: createAccessToken
      summary: Exchange auth code for access + refresh tokens
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Token pair issued.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        "400":
          description: Invalid or missing code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FailResponse"
...